Function App specific checks for the network checker tool  (#1163)
* Initial commit of working code

* Renamed function names to follow the camelCase with first letter lower case convention

* Functions flow is now only shown for Function Apps

* Final Functions checks with multi-level rendering and failure details in sidebar

* An optional docker image for local development

* Fixed git repo name in readme and added instructions to build and run a docker container for local development.

* Minor bug fix

* Addressing PR feedback

* PR feedback

* Scripts and instructions for local development via docker container

* null checks and enabling function app checks on non-vnet joined apps

* Reordered drop-down, changed pending to info in checkstep, new info message when no elegible functions found

* Top level check is info level and added warning as first subcheck
