+++ AngularApp/Dockerfile
@@ -1,0 +1,3 @@
+FROM node:12.22-alpine AS base-layer
+RUN apk update && apk add bash
+RUN npm install
