+++ AngularApp/buildimage.bat
@@ -1,0 +1,1 @@
+docker build -t placeholder.repo:angularappimage .
