+++ AngularApp/buildimage.sh
@@ -1,0 +1,1 @@
+docker build -t placeholder.repo:angularappimage .
