+++ AngularApp/package.json
@@ -8,7 +8,7 @@
     "applens": "node --max_old_space_size=8192 node_modules/@angular/cli/bin/ng serve applens --aot --optimization=false",
     "ncloud-applens-prod": "ng serve applens --configuration=nationalcloud-production --aot --optimization=false",
     "ncloud-applens-local": "ng serve applens --configuration=nationalcloud-local --aot --optimization=false",
-    "ssl": "node --max_old_space_size=8192 node_modules/@angular/cli/bin/ng serve app-service-diagnostics --aot --optimization=false --ssl --sslCert ./ssl/server.crt --sslKey ./ssl/server.key",
+    "ssl": "node --max_old_space_size=8192 node_modules/@angular/cli/bin/ng serve --host 0.0.0.0 app-service-diagnostics --aot --optimization=false --ssl --sslCert ./ssl/server.crt --sslKey ./ssl/server.key",
     "ssl-local": "npm run ssl -- -c local",
     "asd": "concurrently -k \"npm run ssl\" \"npm run lib\"",
     "build-lib": "node --max_old_space_size=8192 node_modules/@angular/cli/bin/ng build diagnostic-data",
