+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/step-views/step-view-lib.ts
@@ -69,7 +69,7 @@ export enum checkResultLevel {
     pass,
     warning,
     fail,
-    pending,
+    info,
     loading,
     error,
     hidden
