+++ AngularApp/startcontainer.bat
@@ -1,0 +1,1 @@
+docker run -it -v %CD%%:/AngularApp -p 0.0.0.0:3000:3000 placeholder.repo:angularappimage bash
