+++ AngularApp/startcontainer.sh
@@ -1,0 +1,1 @@
+docker run -it -v ${PWD}:/AngularApp -p 0.0.0.0:3000:3000 placeholder.repo:angularappimage bash
