+++ README.md
@@ -58,8 +58,10 @@ root
 
 ## Getting Started
 
-- Clone repo `git clone https://github.com/Azure/Azure-WebApps-Support-Center.git`
-- Navigate to the angular root folder `AngularApp` and run `npm install` (This will install all the required packages.)
+- Clone repo `git clone https://github.com/Azure/Azure-AppServices-Diagnostics-Portal.git`
+- Either:
+  - Install required dependencies locally by navigating to the angular root folder `AngularApp` and runing `npm install` (This will install all the required packages.) OR
+  - Build and run a docker container that has all dependencies. Run AngularApp/buildimage.sh (or buildimage.bat on Windows) once to build an image locally and then AngularApp/startcontainer.sh (or startcontaine.bat on Windows) to build a container from that image and open a terminal on it.  Once the container is created, you can use the docker UI to stop/start and open a terminal to the container.
 
 ## App Service Diagnostics Portal
